Moved DependsOnMethod to routing.ts.
