+++ src/index.ts
@@ -1,12 +1,12 @@
 export { ConfigType, LoggerConfig } from './config-type';
 export { AbstractEndpoint, EndpointInput, EndpointOutput } from './endpoint';
-export { Method, DependsOnMethod } from './method';
+export { Method } from './method';
 export { EndpointsFactory } from './endpoints-factory';
 export { IOSchema, FlatObject } from './helpers';
 export { createLogger } from './logger';
 export { createMiddleware } from './middleware';
 export { ResultHandler } from './result-handler';
-export { Routing } from './routing';
+export { Routing, DependsOnMethod } from './routing';
 export { createServer, attachRouting } from './server';
 export { OpenAPI } from './open-api';
 export { OpenAPIError, DependsOnMethodError } from './errors';
