+++ tests/unit/routing.spec.ts
@@ -1,6 +1,6 @@
 import {Express, RequestHandler, Request, Response} from 'express';
 import {Logger} from 'winston';
-import {EndpointsFactory, z, Routing, ConfigType} from '../../src';
+import {EndpointsFactory, z, Routing, ConfigType, DependsOnMethod} from '../../src';
 import {initRouting} from '../../src/routing';
 
 let appMock: any;
@@ -131,6 +131,56 @@ describe('Routing', () => {
           result: true
         }
       });
+    });
+  });
+
+  describe('DependsOnMethod', () => {
+    test('should accept empty object', () => {
+      const instance = new DependsOnMethod({});
+      expect(instance).toBeInstanceOf(DependsOnMethod);
+      expect(instance.methods).toEqual({});
+    });
+
+    test('should accept an endpoint with a corresponding method', () => {
+      const instance = new DependsOnMethod({
+        post: new EndpointsFactory().build({
+          method: 'post',
+          input: z.object({}),
+          output: z.object({}),
+          handler: async () => ({})
+        })
+      });
+      expect(instance).toBeInstanceOf(DependsOnMethod);
+      expect(instance.methods).toHaveProperty('post');
+    });
+
+    test('should accept an endpoint with additional methods', () => {
+      const endpoint = new EndpointsFactory().build({
+        methods: ['get', 'post'],
+        input: z.object({}),
+        output: z.object({}),
+        handler: async () => ({})
+      });
+      const instance = new DependsOnMethod({
+        get: endpoint,
+        post: endpoint
+      });
+      expect(instance).toBeInstanceOf(DependsOnMethod);
+      expect(instance.methods).toHaveProperty('get');
+      expect(instance.methods).toHaveProperty('post');
+    });
+
+    test('should throw an error if the endpoint does not have the corresponding method', () => {
+      const endpoint = new EndpointsFactory().build({
+        methods: ['get', 'patch'],
+        input: z.object({}),
+        output: z.object({}),
+        handler: async () => ({})
+      });
+      expect(() => new DependsOnMethod({
+        get: endpoint,
+        post: endpoint
+      })).toThrowErrorMatchingSnapshot();
     });
   });
 });
